# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Panier API Integration",
        "description": "API endpoints for communicating with the Panier API - ZIMRA Fiscalisation Support.\r\n\r\n## Rate Limit\r\n500 API calls every 5 minutes for each unique IP Address.\r\n\r\n## Authentication\r\nRequires APP-ID and API-KEY headers obtained from your Panier company account's Developer API section.\r\n\r\n## ZIMRA Fiscalisation\r\nThis API supports ZIMRA Fiscal Invoices, Credit Notes and Debit Notes based on ZIMRA Fiscal Device Gateway API Specs v7.2.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/sale/create": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Create a sale",
                "description": "Create a sale with optional ZIMRA fiscalization",
                "operationId": "14efb4baa3a5ce508835758c102b3b9f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "customer_id": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "currency_id": {
                                                "type": "string"
                                            },
                                            "products": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object"
                                                }
                                            },
                                            "payment_method": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "zimra_fiscalize": {
                                        "type": "boolean",
                                        "default": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/sale/search": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Search sales",
                "operationId": "5b90bcddd7b6028c5b4d8f066d6904ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "query": {
                                                "type": "string",
                                                "default": "*"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/sale/void": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Void a sale",
                "operationId": "5e97ed23e1d84736eb423cc001dc28cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "reason": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/sale/download": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Download sale receipt as PDF",
                "operationId": "45ffdc5077ef1fa5e23de4b6d041e793",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/payment-provider/check-payment-status": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Check payment status",
                "operationId": "90d553532917e7ac1c799776679ee750",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment status"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/payment-provider/confirm-payment": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Confirm payment",
                "operationId": "d493195ee9a7c6634c4f6ebdf4352a11",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "AppId": {
                "type": "apiKey",
                "description": "Panier APP-ID from Developer API section",
                "name": "APP-ID",
                "in": "header"
            },
            "ApiKey": {
                "type": "apiKey",
                "description": "Panier API-KEY from Developer API section",
                "name": "API-KEY",
                "in": "header"
            }
        }
    },
...
}